Implemented PauseEvent and made ToJson in JsonConversionService not abstract
