+++ Hunted Mobile/Hunted Mobile/Hunted Mobile/Service/Json/LootJsonService.cs
@@ -7,10 +7,6 @@ using System.Text;
 
 namespace Hunted_Mobile.Service.Json {
     public class LootJsonService : JsonConversionService<Loot, LootData> {
-        public override string ToJson(Loot @object) {
-            throw new NotImplementedException();
-        }
-
         protected override Loot ToObject(LootData data) {
             return new Loot(data.id) {
                 Location = new LocationJsonService().ToObjectFromCsv(data.location),
